Rename clawd-code examples to moltbot
Rename clawdbot-aws-typescript and clawdbot-hetzner-typescript examples
to moltbot-aws-typescript and moltbot-hetzner-typescript to reflect the
new product naming. Update package names and project configurations.
