--- clawd-code/clawdbot-aws-typescript/Pulumi.yaml
@@ -1,3 +1,0 @@
-name: clawdbot-aws-typescript
-runtime: nodejs
-description: Deploy Clawdbot AI assistant to AWS with VPC, security groups, and EC2 instance using a reusable ComponentResource.
