--- clawd-code/clawdbot-hetzner-typescript/Pulumi.yaml
@@ -1,3 +1,0 @@
-name: clawdbot-hetzner-typescript
-runtime: nodejs
-description: Deploy Clawdbot AI assistant to Hetzner Cloud with firewall and server using a reusable ComponentResource.
