+++ moltbot/moltbot-hetzner-typescript/Pulumi.yaml
@@ -1,0 +1,3 @@
+name: moltbot-hetzner-typescript
+runtime: nodejs
+description: Deploy Moltbot AI assistant to Hetzner Cloud with firewall and server using a reusable ComponentResource.
