+++ moltbot/moltbot-hetzner-typescript/tsconfig.json
@@ -12,5 +12,5 @@
         "noImplicitReturns": true,
         "forceConsistentCasingInFileNames": true
     },
-    "files": ["index.ts", "clawdbot.ts"]
+    "files": ["index.ts"]
 }
